requirements/test.txt
-r main.txt
nose
mock
coverage
django-nose
nose-cov
git+https://github.com/alisaifee/pymemcache.git#egg=pymemcache
redis
hiro>0.1.6
ordereddict
